Add SMTs to create Json String from String Array and Struct
There is no built-in SMT to handle converting a String array or a Struct
to a Field that can be inserted in to a Database column.  This change
adds two SMTs that will transform a String array or a Struct Field from
a Struct to a JSON String value which can be mapped to a jsonb column in
Postgres for example.

Each SMT has validators attached to the configuration fields to ensure
the SMT is configured correctly before it is created.  Connector config
validation details can be found in the page, and the same validation
semantics apply to the SMT configuration+
- https://www.confluent.io/blog/write-a-kafka-connect-connector-with-configuration-handling/#validators

A CompositeValidator with a NonNullValidator/NonEmptyString is used to
ensure a valid value for the mandatory configuration field is specified
for each SMT.

See also:
- https://cwiki.apache.org/confluence/display/KAFKA/KIP-66%3A+Single+Message+Transforms+for+Kafka+Connect
- https://cwiki.apache.org/confluence/display/KAFKA/KIP-802%3A+Validation+Support+for+Kafka+Connect+SMT+and+Converter+Options