Add JdbcSinkConnector connector, create at startup
When starting a new Kafka Connect Worker instance, it is useful for it
to contain all the connector configurations that it should be running.

This change adds support for creating multiple connectors when the
instance starts, after it begins accepting requests from the REST API
at `http://localhost:8083/connectors`.

The connector config JSON files are copied to the Docker image, and a
new entrypoint script at `/connect-scripts/connect-entrypoint.sh` will:
1. start up the service using the bundled confluent scripts
2. wait for the REST API to be ready to accept requests
3. and then make a http PUT call to create/update each of the connectors
   it finds a config for in the `/connector-configs` directory

See more info on creating connectors at startup in this guide
 - https://developer.confluent.io/courses/kafka-connect/docker-containers/#add-connector-instance-at-container-launch

---

An AVRO schema and sample data has been added so that we can manually
produce data to the topic that the connector consumes from.
 - *In future this will be performed in an automated Integration test*

---

Other changes
- Install `jq` to extract connector name from connector config json file
- Copy the jar file with custom SMTs that are used by the connector
- Make the connect service depend on the postgres service so that it
  will be available and data can be sent from the connector