Add JaCoCo Report Aggregation subproject
The JaCoCo Report Aggregation Plugin allows us to get coverage metrics
of the code in all subprojects in a single HTML/xml report, which can be
used by Sonar or a CI tool to ensure quality gates are being met.

A new `jacocoFullReport` Gradle Task has been added to generate
the aggregate report.  After running the `check` or `jacocoFullReport`
Task you can find the aggregate report in the following directory:
- `jacoco-report-aggregation/build/reports/jacoco/`

The execution data used to build the report is manually searched for in
each subproject:
- `executionData fileTree(project.rootDir.absolutePath).include("**/build/jacoco/*.exec")`

We can add more subprojects and their coverage data will be appended
automatically, we just add a `dependsOn` entry for the new subprojects.

---

Other Changes:
1. Move all plugin definitions to a build file in the project root to
   centralise version numbers. Subprojects can apply the plugins if they
   need them (the `connect-spring-boot-app` and
   `jacoco-report-aggregation` subprojects both need the spring boot
   plugins applied).
2. Added a `bevans.spring-boot-conventions` plugin to centralise the
   custom repo that provides the `kafka-avro-serializer` dependency.

---

See more at:
- https://docs.gradle.org/current/userguide/jacoco_report_aggregation_plugin.html