Add Spring Boot App integration test subproject
A Spring Boot application has been added which provides two main
features to assist with running end to end integration tests:
1. Producing a Kafka Record with an AVRO payload Value to a topic
2. Fetching an Entity from a JPA Repository

The spring boot test assumes that a JDBCSinkConnector is deployed to the
Kafka Connect `connect` container in the Docker Compose stack that will
read the data from the topic in (1) above and insert into the Database
table so that it can be read from (2) above using Spring Data JPA.

The Docker Compose project used to run the integration tests is named
`kafka-connect-stack_inttest` by the `com.avast.gradle.docker-compose`
gradle plugin (via the `intTest` nested configuration).

Running `./gradlew check` will run the `integrationTest` task which will
bring the Docker Compose stack up, run the tests, bring it down, and
remove all the containers.

It is possible to run the integration tests without bringing the
Docker Compose stack down (and the containers being removed).  This is
useful if you are making changes to the tests, the underlying Spring
Boot application, or even the SMTs in the `connect-smt-lib` library
(making changes to the latter will force the `connect` container to be
recreated, which is desirable).

To do this run `./gradlew integrationTestRun`.

An `integrationTest` sourceSet has been added so that we can separate
the unit tests and the integration tests, the former running in a much
shorter time by running the standard `./gradlew test` task.

---

The main gradle plugins being used are:
1. `com.avast.gradle.docker-compose` - to interact with Docker Compose
2. `com.github.davidmc24.gradle.plugin.avro` - to generate POJOs from
                                               AVRO schema files.

---

Dependencies to note are:
1. `spring-kafka` - to send Records to Kafka using a `KafkaTemplate`
2. `spring-boot-starter-data-jpa` - to use JPA to interact with Postgres
3. `spring-boot-starter-test`     - to get all the testing goodies
   - https://docs.spring.io/spring-boot/docs/3.2.4/reference/html/features.html#features.testing.test-scope-dependencies

Watch this Confluent course to learn about Spring Boot for Apache Kafka:
 - https://developer.confluent.io/courses/spring/apache-kafka-intro/

---

Other changes
1. Comment out `depends_on` connect service so that we can recreate
   connect without having to recreate the dependant Docker Compose
   services such as control center etc.

2. Jacoco configs have been added to help get code coverage from both
   the unit tests and the integration tests.

---

See more at
- https://github.com/avast/gradle-docker-compose-plugin/
- https://ahassany.medium.com/automatically-launch-docker-compose-from-gradle-for-dev-and-integration-tests-4a63b78ad7a1
- https://docs.gradle.org/current/userguide/java_testing.html#sec:configuring_java_integration_tests

- https://docs.spring.io/spring-kafka/docs/3.1.4/reference/kafka/sending-messages.html
- https://docs.spring.io/spring-boot/docs/3.2.4/reference/htmlsingle/#features.testing.spring-boot-applications
- https://developer.confluent.io/courses/spring/apache-kafka-intro/

- https://github.com/davidmc24/gradle-avro-plugin