Add project purpose etc. to the README
In order to give the reader a better understanding of the purpose of the
project the README has been updated to include details of the main
technologies used such as the various Gradle plugins, Docker, Confluent
Platform, Kafka Connect, and JaCoCo.

The project contains several modules to improve separation of code and
details about each one has been added to the README.

Some Gradle commands have been provided to assist in running the tests
after checking out the repository from GitHub.

The reader should have a good idea of when to expect to learn by
reviewing README, the code, and running the integration tests.