Add JDWP Debugging config to Connect container
When developing SMTs it is very useful to be able to debug the code in
a running Docker container, in order to do this we need to enable the
jdwp protocol in the running JVM.

You can do this by adding the following environment variable to the
container in the docker compose service and exposing the port:
- `JAVA_TOOL_OPTIONS: "-agentlib:jdwp=transport=dt_socket,server=y,suspend=n,address=*:8000"`

Please note "3.3. Since Java 9" we need to add an asterix to be able to
connect to the remote port in the container from our IDE.  According to
the documentation, this is not secure, and it’s recommended to specify
the debugger’s IP address whenever possible.

---

See more details at:
- https://www.baeldung.com/java-application-remote-debugging
- https://medium.com/swlh/remote-debugging-a-java-application-running-in-docker-container-with-intellij-idea-efe54cd77f02