Make `integrationTest` task sleep for 10 seconds at start (#18)
When a second integration test was added it was intermittently failing
due to the Connect service not having enough time to finish creating all
the connectors at startup time before tests started running. This
resulted in the data not being present in the database which the
JDBCSinkConnector inserts.

A sleep delay has been added in the `doFirst` sub-task so that the
Connect container has enough time to auto create all the connectors at
startup before any of the tests run.  If we add a lot more integration
tests this delay may need to be increased.

The delay between creating connectors has also been decreased from ten
to two seconds, and the delay between the DELETE and PUT rest api calls
is decreased from five to two seconds.