Enable error logging and add details to deadletterqueue headers
When an error occurs in a connector task, the offending record is sent
to the Dead Letter Queue but no information about the error is present.

This change adds the information about the reason for a message’s
rejection into the header of the message itself, including the
stacktrace which is usually only available in the log file.

```json
{
  ...
  "errors.log.enable": "true",
  "errors.log.include.messages": "true",
  "errors.deadletterqueue.context.headers.enable": "true"
  ...
}
```

See more at:
- https://www.confluent.io/blog/kafka-connect-deep-dive-error-handling-dead-letter-queues/#b7488c74-e081-4866-bb87-9ec0086bea5e

- https://docs.confluent.io/platform/current/installation/configuration/connect/sink-connect-configs.html#errors-deadletterqueue-context-headers-enable