Add FF VII Battle update JDBCSourceConnector x 3
Battle data for the FF VII game is stored in a MySQL database which is
produced by a legacy system. We want to be able to source data from the
database and add it to a Kafka Topic, which can then be consumed by
another Kafka application, most likely a JDBCSinkConnector.
 - https://docs.confluent.io/kafka-connectors/jdbc/current/source-connector/overview.html

A connector config for each of the three different incremental query
modes have been added and each one has a unique target topic due to its
prefix:
- https://docs.confluent.io/kafka-connectors/jdbc/current/source-connector/overview.html#incremental-query-modes

---

The mysql-connector-j-8.4.0.jar connector needs to be manually
downloaded as it is not included in the JDBCSourceConnector plugin due
to Oracle licensing issues.

Init scripts for the newly added MySQL container have also been added.