Add healthcheck to broker/connect/schema registry
The broker, connect, and schema registry services have been known to
fail; without health checks it is difficult to tell if they are working.

The healthcheck attribute declares a check that's run to determine
whether the service containers are "healthy". It works in the same way,
and has the same default values, as the HEALTHCHECK Dockerfile
instruction set by the service's Docker image. Your Compose file can
override the values set in the Dockerfile.
 - https://docs.docker.com/compose/compose-file/05-services/#healthcheck

For more information on HEALTHCHECK, see the Dockerfile reference.
- https://docs.docker.com/reference/dockerfile/#healthcheck