Use Spring auto configuration for Kafka Producer (#39)
Rather the manually creating a `ProducerFactory` and `KafkaTemplate`
and applying the serializer and schema registry url, we can use
properties in the `application.yml` file so that they are picked up by
Spring auto configuration.

All additional properties under `spring.kafka.properties` will be
applied to both consumers and producers. The main change supporting auto
configuration is `spring.kafka.properties.schema.registry.url`.

The current value serializer is AVRO. If we want to use a KafkaProducer
that produces a value that is not AVRO, we would need to reinstate the
custom `ProducerFactory` & `KafkaTemplate` beans to support the other
data type (e.g. JSON) for the value.

See more at
- https://docs.spring.io/spring-boot/reference/messaging/kafka.html#messaging.kafka.additional-properties
- https://docs.spring.io/spring-boot/appendix/application-properties/index.html#application-properties.integration.spring.kafka.properties
- https://docs.spring.io/spring-kafka/docs/3.1.3/reference/kafka/sending-messages.html#examples