Refactor @Resource content loading to a @Component
The ability to get a String from a Spring `@Resource` is logic that can
be reused, and it should be encapsulated in a utility `@Component`.

This change adds a class that will provide a `getStringFromResource`
method to load the content of a resource and convert it to a String.  If
the resource cannot be located, a custom checked `ContentLoadException`
will be thrown.

This logic already existed but is now refactored in to a new class.